# perf(core): improve the internal Worker pool